requirements.txt
click==5.1
colorama==0.3.3
pyechonest==9.0.0
PyYAML==3.11
requests==2.8.1
spotipy==2.3.7
tabulate==0.7.5
wheel==0.24.0
